+++ rpc/connector/connector_grpc.pb.go
@@ -76,6 +76,8 @@ type ConnectorClient interface {
 	ListCommands(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*CommandGroups, error)
 	// RunCommand executes a CLI command.
 	RunCommand(ctx context.Context, in *RunCommandRequest, opts ...grpc.CallOption) (*RunCommandResponse, error)
+	// ValidArgsForCommand handles autocompletion
+	ValidArgsForCommand(ctx context.Context, in *ValidArgsForCommandRequest, opts ...grpc.CallOption) (*ValidArgsForCommandResponse, error)
 	// ResolveIngressInfo is a temporary rpc intended to allow the cli to ask
 	// the cloud for default ingress values
 	ResolveIngressInfo(ctx context.Context, in *userdaemon.IngressInfoRequest, opts ...grpc.CallOption) (*userdaemon.IngressInfoResponse, error)
@@ -87,6 +89,8 @@ type ConnectorClient interface {
 	AddInterceptor(ctx context.Context, in *Interceptor, opts ...grpc.CallOption) (*emptypb.Empty, error)
 	// RemoveInterceptor removes a previously added interceptor
 	RemoveInterceptor(ctx context.Context, in *Interceptor, opts ...grpc.CallOption) (*emptypb.Empty, error)
+	// GetNamespaces gets the mapped namespaces with an optional prefix
+	GetNamespaces(ctx context.Context, in *GetNamespacesRequest, opts ...grpc.CallOption) (*GetNamespacesResponse, error)
 }
 
 type connectorClient struct {
@@ -341,6 +345,15 @@ func (c *connectorClient) RunCommand(ctx context.Context, in *RunCommandRequest,
 	return out, nil
 }
 
+func (c *connectorClient) ValidArgsForCommand(ctx context.Context, in *ValidArgsForCommandRequest, opts ...grpc.CallOption) (*ValidArgsForCommandResponse, error) {
+	out := new(ValidArgsForCommandResponse)
+	err := c.cc.Invoke(ctx, "/telepresence.connector.Connector/ValidArgsForCommand", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 func (c *connectorClient) ResolveIngressInfo(ctx context.Context, in *userdaemon.IngressInfoRequest, opts ...grpc.CallOption) (*userdaemon.IngressInfoResponse, error) {
 	out := new(userdaemon.IngressInfoResponse)
 	err := c.cc.Invoke(ctx, "/telepresence.connector.Connector/ResolveIngressInfo", in, out, opts...)
@@ -371,6 +384,15 @@ func (c *connectorClient) AddInterceptor(ctx context.Context, in *Interceptor, o
 func (c *connectorClient) RemoveInterceptor(ctx context.Context, in *Interceptor, opts ...grpc.CallOption) (*emptypb.Empty, error) {
 	out := new(emptypb.Empty)
 	err := c.cc.Invoke(ctx, "/telepresence.connector.Connector/RemoveInterceptor", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *connectorClient) GetNamespaces(ctx context.Context, in *GetNamespacesRequest, opts ...grpc.CallOption) (*GetNamespacesResponse, error) {
+	out := new(GetNamespacesResponse)
+	err := c.cc.Invoke(ctx, "/telepresence.connector.Connector/GetNamespaces", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -431,6 +453,8 @@ type ConnectorServer interface {
 	ListCommands(context.Context, *emptypb.Empty) (*CommandGroups, error)
 	// RunCommand executes a CLI command.
 	RunCommand(context.Context, *RunCommandRequest) (*RunCommandResponse, error)
+	// ValidArgsForCommand handles autocompletion
+	ValidArgsForCommand(context.Context, *ValidArgsForCommandRequest) (*ValidArgsForCommandResponse, error)
 	// ResolveIngressInfo is a temporary rpc intended to allow the cli to ask
 	// the cloud for default ingress values
 	ResolveIngressInfo(context.Context, *userdaemon.IngressInfoRequest) (*userdaemon.IngressInfoResponse, error)
@@ -442,6 +466,8 @@ type ConnectorServer interface {
 	AddInterceptor(context.Context, *Interceptor) (*emptypb.Empty, error)
 	// RemoveInterceptor removes a previously added interceptor
 	RemoveInterceptor(context.Context, *Interceptor) (*emptypb.Empty, error)
+	// GetNamespaces gets the mapped namespaces with an optional prefix
+	GetNamespaces(context.Context, *GetNamespacesRequest) (*GetNamespacesResponse, error)
 	mustEmbedUnimplementedConnectorServer()
 }
 
@@ -515,6 +541,9 @@ func (UnimplementedConnectorServer) ListCommands(context.Context, *emptypb.Empty
 func (UnimplementedConnectorServer) RunCommand(context.Context, *RunCommandRequest) (*RunCommandResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method RunCommand not implemented")
 }
+func (UnimplementedConnectorServer) ValidArgsForCommand(context.Context, *ValidArgsForCommandRequest) (*ValidArgsForCommandResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ValidArgsForCommand not implemented")
+}
 func (UnimplementedConnectorServer) ResolveIngressInfo(context.Context, *userdaemon.IngressInfoRequest) (*userdaemon.IngressInfoResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ResolveIngressInfo not implemented")
 }
@@ -526,6 +555,9 @@ func (UnimplementedConnectorServer) AddInterceptor(context.Context, *Interceptor
 }
 func (UnimplementedConnectorServer) RemoveInterceptor(context.Context, *Interceptor) (*emptypb.Empty, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method RemoveInterceptor not implemented")
+}
+func (UnimplementedConnectorServer) GetNamespaces(context.Context, *GetNamespacesRequest) (*GetNamespacesResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetNamespaces not implemented")
 }
 func (UnimplementedConnectorServer) mustEmbedUnimplementedConnectorServer() {}
 
@@ -942,6 +974,24 @@ func _Connector_RunCommand_Handler(srv interface{}, ctx context.Context, dec fun
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Connector_ValidArgsForCommand_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ValidArgsForCommandRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ConnectorServer).ValidArgsForCommand(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/telepresence.connector.Connector/ValidArgsForCommand",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ConnectorServer).ValidArgsForCommand(ctx, req.(*ValidArgsForCommandRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _Connector_ResolveIngressInfo_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(userdaemon.IngressInfoRequest)
 	if err := dec(in); err != nil {
@@ -1010,6 +1060,24 @@ func _Connector_RemoveInterceptor_Handler(srv interface{}, ctx context.Context, 
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ConnectorServer).RemoveInterceptor(ctx, req.(*Interceptor))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Connector_GetNamespaces_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetNamespacesRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ConnectorServer).GetNamespaces(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/telepresence.connector.Connector/GetNamespaces",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ConnectorServer).GetNamespaces(ctx, req.(*GetNamespacesRequest))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -1102,6 +1170,10 @@ var Connector_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _Connector_RunCommand_Handler,
 		},
 		{
+			MethodName: "ValidArgsForCommand",
+			Handler:    _Connector_ValidArgsForCommand_Handler,
+		},
+		{
 			MethodName: "ResolveIngressInfo",
 			Handler:    _Connector_ResolveIngressInfo_Handler,
 		},
@@ -1116,6 +1188,10 @@ var Connector_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "RemoveInterceptor",
 			Handler:    _Connector_RemoveInterceptor_Handler,
+		},
+		{
+			MethodName: "GetNamespaces",
+			Handler:    _Connector_GetNamespaces_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{